Crate squash_sys

Structs

SquashCodec

SquashCodecImpl

SquashCodecInfo

SquashContext

SquashFile

SquashLicense

SquashMemoryFuncs

SquashObject

SquashOptionInfo

SquashOptionInfoEnumInt

SquashOptionInfoEnumString

SquashOptionInfoEnumStringMap

SquashOptionInfoRangeInt

SquashOptionInfoRangeSize

SquashOptions

SquashPlugin

SquashStream

SquashStreamPrivate

Enums

FILE

Constants

SQUASH_BAD_PARAM

SQUASH_BAD_VALUE

SQUASH_BUFFER_EMPTY

SQUASH_BUFFER_FULL

SQUASH_CODEC_INFO_AUTO_MASK

SQUASH_CODEC_INFO_CAN_FLUSH

SQUASH_CODEC_INFO_DECOMPRESS_UNSAFE

SQUASH_CODEC_INFO_KNOWS_UNCOMPRESSED_SIZE

SQUASH_CODEC_INFO_MASK

SQUASH_CODEC_INFO_NATIVE_STREAMING

SQUASH_CODEC_INFO_VALID

SQUASH_CODEC_INFO_WRAP_SIZE

SQUASH_END_OF_STREAM

SQUASH_FAILED

SQUASH_INVALID_BUFFER

SQUASH_INVALID_OPERATION

SQUASH_IO

SQUASH_LICENSE_APACHE

SQUASH_LICENSE_APACHE2

SQUASH_LICENSE_BSD2

SQUASH_LICENSE_BSD3

SQUASH_LICENSE_BSD4

SQUASH_LICENSE_CDDL

SQUASH_LICENSE_COPYLEFT_INCOMPATIBLE

SQUASH_LICENSE_FLAGS_MASK

SQUASH_LICENSE_GPL1

SQUASH_LICENSE_GPL2

SQUASH_LICENSE_GPL3

SQUASH_LICENSE_GPL1_PLUS

SQUASH_LICENSE_GPL2_PLUS

SQUASH_LICENSE_GPL3_PLUS

SQUASH_LICENSE_ISC

SQUASH_LICENSE_LGPL3

SQUASH_LICENSE_LGPL2P1

SQUASH_LICENSE_LGPL2P1_PLUS

SQUASH_LICENSE_LGPL3_PLUS

SQUASH_LICENSE_MIT

SQUASH_LICENSE_MPL

SQUASH_LICENSE_MSPL

SQUASH_LICENSE_OR_GREATER

SQUASH_LICENSE_PERMISSIVE

SQUASH_LICENSE_PROPRIETARY

SQUASH_LICENSE_PUBLIC_DOMAIN

SQUASH_LICENSE_STRONG_COPYLEFT

SQUASH_LICENSE_TYPE_MASK

SQUASH_LICENSE_UNKNOWN

SQUASH_LICENSE_WEAK_COPYLEFT

SQUASH_LICENSE_WTFPL

SQUASH_LICENSE_X11

SQUASH_LICENSE_ZLIB

SQUASH_MEMORY

SQUASH_NOT_FOUND

SQUASH_OK

SQUASH_OPERATION_FINISH

SQUASH_OPERATION_FLUSH

SQUASH_OPERATION_PROCESS

SQUASH_OPERATION_TERMINATE

SQUASH_OPTION_TYPE_BOOL

SQUASH_OPTION_TYPE_ENUM_INT

SQUASH_OPTION_TYPE_ENUM_STRING

SQUASH_OPTION_TYPE_INT

SQUASH_OPTION_TYPE_NONE

SQUASH_OPTION_TYPE_RANGE_INT

SQUASH_OPTION_TYPE_RANGE_SIZE

SQUASH_OPTION_TYPE_SIZE

SQUASH_OPTION_TYPE_STRING

SQUASH_PROCESSING

SQUASH_RANGE

SQUASH_STATE

SQUASH_STREAM_COMPRESS

SQUASH_STREAM_DECOMPRESS

SQUASH_STREAM_STATE_FINISHED

SQUASH_STREAM_STATE_FINISHING

SQUASH_STREAM_STATE_FLUSHING

SQUASH_STREAM_STATE_IDLE

SQUASH_STREAM_STATE_RUNNING

SQUASH_UNABLE_TO_LOAD

SQUASH_VERSION_API
API version

SQUASH_VERSION_CURRENT
Current version, encoded as a single number

SQUASH_VERSION_MAJOR
Major version number

SQUASH_VERSION_MINOR
Minor version number

SQUASH_VERSION_REVISION
Revsion version number

Functions

SQUASH_VERSION
Encode the major, minor, and revisions into a single number

SQUASH_VERSION_EXTRACT_MAJOR
Extract the major version number from an encoded version

SQUASH_VERSION_EXTRACT_MINOR
Extract the minor version number from an encoded version

SQUASH_VERSION_EXTRACT_REVISION
Extract the revision number from an encoded version

squash_aligned_alloc⚠

squash_aligned_free⚠

squash_calloc⚠

squash_codec_compress⚠

squash_codec_compress_with_options⚠

squash_codec_create_stream⚠

squash_codec_create_stream_with_options⚠

squash_codec_decompress⚠

squash_codec_decompress_with_options⚠

squash_codec_get_context⚠

squash_codec_get_extension⚠

squash_codec_get_info⚠

squash_codec_get_max_compressed_size⚠

squash_codec_get_name⚠

squash_codec_get_option_info⚠

squash_codec_get_plugin⚠

squash_codec_get_priority⚠

squash_codec_get_uncompressed_size⚠

squash_codec_init⚠

squash_context_foreach_codec⚠

squash_context_foreach_plugin⚠

squash_context_get_codec⚠

squash_context_get_codec_from_extension⚠

squash_context_get_default⚠

squash_context_get_plugin⚠

squash_error⚠

squash_file_close⚠

squash_file_eof⚠

squash_file_error⚠

squash_file_flush⚠

squash_file_flush_unlocked⚠

squash_file_free⚠

squash_file_lock⚠

squash_file_open⚠

squash_file_open_with_options⚠

squash_file_printf⚠

squash_file_read⚠

squash_file_read_unlocked⚠

squash_file_steal⚠

squash_file_steal_with_options⚠

squash_file_unlock⚠

squash_file_wopen⚠

squash_file_wopen_with_options⚠

squash_file_wprintf⚠

squash_file_write⚠

squash_file_write_unlocked⚠

squash_foreach_codec⚠

squash_foreach_plugin⚠

squash_free⚠

squash_get_codec⚠

squash_get_codec_from_extension⚠

squash_get_plugin⚠

squash_license_from_string⚠

squash_license_to_string⚠

squash_malloc⚠

squash_object_destroy⚠

squash_object_get_ref_count⚠

squash_object_init⚠

squash_object_ref⚠

squash_object_ref_sink⚠

squash_object_unref⚠

squash_options_destroy⚠

squash_options_get_bool⚠

squash_options_get_bool_at⚠

squash_options_get_boolw⚠

squash_options_get_int⚠

squash_options_get_int_at⚠

squash_options_get_intw⚠

squash_options_get_size⚠

squash_options_get_size_at⚠

squash_options_get_sizew⚠

squash_options_get_string⚠

squash_options_get_string_at⚠

squash_options_get_stringw⚠

squash_options_init⚠

squash_options_new⚠

squash_options_newa⚠

squash_options_newaw⚠

squash_options_neww⚠

squash_options_parse⚠

squash_options_parse_option⚠

squash_options_parse_optionw⚠

squash_options_parsea⚠

squash_options_parseaw⚠

squash_options_parsew⚠

squash_options_set_bool⚠

squash_options_set_bool_at⚠

squash_options_set_int⚠

squash_options_set_int_at⚠

squash_options_set_size⚠

squash_options_set_size_at⚠

squash_options_set_string⚠

squash_options_set_string_at⚠

squash_plugin_foreach_codec⚠

squash_plugin_get_codec⚠

squash_plugin_get_licenses⚠

squash_plugin_get_name⚠

squash_plugin_init⚠

squash_realloc⚠

squash_set_default_search_path⚠

squash_set_memory_functions⚠

squash_splice⚠

squash_splice_custom⚠

squash_splice_custom_with_options⚠

squash_splice_with_options⚠

squash_status_to_string⚠

squash_stream_destroy⚠

squash_stream_finish⚠

squash_stream_flush⚠

squash_stream_init⚠

squash_stream_new⚠

squash_stream_new_with_options⚠

squash_stream_newa⚠

squash_stream_process⚠

squash_version⚠

squash_version_api⚠

Type Definitions

SquashCodecForeachFunc

SquashDestroyNotify

SquashOperation

SquashOptionType

SquashPluginForeachFunc

SquashReadFunc

SquashStatus

SquashStreamState

SquashStreamType

SquashWriteFunc

wchar_t

Unions

SquashOptionInfo__bindgen_ty_1

SquashOptionValue