squash_sys::squash_aligned_alloc

pub unsafe extern fn squash_aligned_alloc(alignment: usize, size: usize) -> *mut c_void

Allocate an aligned buffer

Memory allocated with this function is assumed not to support reallocation. In reality, assuming nobody has installed thick wrappers it should be possible to squash_realloc the buffer, but the result is not constrained to the alignment requirements presented to the initial buffer.

The value returned by this function must be freed with squash_aligned_free; While some implementations (such as C11's aligned_alloc and the POSIX posix_memalign function) allow values returned by squash_aligned_alloc to be passed directly to squash_free, others (such as Windows' _aligned_malloc) do not. Passing the result of this function to squash_free is considered undefined behavior.

Note

Values supported for the alignment parameter are implementation defined, but a fair assumption is that they must be a power of two and multiple of sizeof(void*).

Parameters

ctx The context

alignment Alignment of the buffer

size Number of bytes to allocate