squash_sys::squash_codec_init

pub unsafe extern fn squash_codec_init(codec: *mut SquashCodec) -> SquashStatus

Initialize a codec.

Note

This function is generally only useful inside of a callback passed to squash_foreach_codec or squash_plugin_foreach_codec. Every other way to get a codec (such as squash_get_codec or squash_plugin_get_codec) will initialize the codec as well (and return NULL instead of the codec if initialization fails). The foreach functions, however, do not initialize the codec since doing so requires actually loading the plugin.

Parameters

codec The codec.

Returns

A status code.

Return values

SQUASH_OK Codec successfully initialized.

SQUASH_UNABLE_TO_LOAD Failed to load the codec.