squash_sys::squash_file_close

pub unsafe extern fn squash_file_close(file: *mut SquashFile) -> SquashStatus

Close a file.

If file is a compressor the stream will finish compressing, writing any buffered data. For codecs which do not provide a native streaming interface, all of the actual compression will take place during this call. In other words, it may block for a non-trivial period. If this is a problem please file a bug against Squash (including your use case), and we can discuss adding a function call which will simply abort compression.

In addition to freeing the SquashFile instance, this function will close the underlying FILE pointer. If you wish to continue using the FILE for something else, use squash_file_free instead.

Parameters

file file to close

Returns

SQUASH_OK on success or a negative error code on failure