squash_sys::squash_file_flush

pub unsafe extern fn squash_file_flush(file: *mut SquashFile) -> SquashStatus

Immediately write any buffered data to a file.

Note

This function only works for codecs which support flushing (see the SQUASH_CODEC_INFO_CAN_FLUSH flag in the return value of squash_codec_get_info).

Parameters

file file to flush

Returns

true if flushing succeeeded, false if flushing is not supported or there was another error.