squash_sys::squash_file_flush_unlocked

pub unsafe extern fn squash_file_flush_unlocked(file: *mut SquashFile) -> SquashStatus

Immediately write any buffered data to a file without acquiring the lock

Write data to a compressed file without acquiring the lock

This function is the same as squash_file_write, except it will not acquire a lock on the SquashFile instance. It should be used only when there is no possibility of other threads accessing the file, or if you have already acquired the lock with squash_file_lock.

Parameters

file file to flush

Returns

true if flushing succeeeded, false if flushing is not supported or there was another error.