squash_sys::squash_file_lock

pub unsafe extern fn squash_file_lock(file: *mut SquashFile)

Acquire the lock on a file.

squash_file_read, squash_file_write, and squash_file_flush are thread-safe. This is accomplished by acquiring a lock on while each function is operating in order to ensure exclusive access.

If, however, the programmer wishes to call a series of functions and ensure that they are performed without interference, they can manually acquire the lock with this function and use the unlocked variants (squash_file_read_unlocked, squash_file_write_unlocked, and squash_file_flush_unlocked).

Note

This function has nothing to do with the kind of lock acquired by the flock function.

Parameters

file the file to acquire the lock on