squash_sys::squash_file_open

pub unsafe extern fn squash_file_open(codec: *mut SquashCodec, filename: *const c_char, mode: *const c_char, ...) -> *mut SquashFile

Open a file.

The mode parameter will be passed through to fopen, so the value must valid. Note that Squash may attempt to use mmap regardless of whether the m flag is passed.

The file is always assumed to be compressed—calling squash_file_write will always compress, and calling squash_file_read will always decompress. Note, however, that you cannot mix reading and writing to the same file as you can with a standard FILE.

Note

Error handling for this function is somewhat limited, and it may be difficult to determine the exact nature of problems such as an invalid codec, where errno is not set. If this is unacceptable you should call squash_options_parse yourself and pass the results to squash_file_open_with_options (which will only fail due to the underlying fopen failing).

Parameters

filename name of the file to open

mode file mode

codec codec to use

... options

Returns

The opened file, or NULL on error