squash_sys::squash_file_open_with_options

pub unsafe extern fn squash_file_open_with_options(codec: *mut SquashCodec, filename: *const c_char, mode: *const c_char, options: *mut SquashOptions) -> *mut SquashFile

Open a file using a with the specified options.

filename is assumed to be UTF-8 encoded. On Windows, this function will call squash_file_wopen_with_options internally. On other platforms, filenames on disk are assumed to be in UTF-8 format, therefore the filename is passed through to fopen without any conversion.

Parameters

filename name of the file to open

mode file mode

codec codec to use

options options

Returns

The opened file, or NULL on error