squash_sys::squash_file_read

pub unsafe extern fn squash_file_read(file: *mut SquashFile, decompressed_size: *mut size_t, decompressed: *mut uint8_t) -> SquashStatus

Read from a compressed file.

Attempt to read decompressed_size bytes of decompressed data into the decompressed buffer. The number of bytes of compressed data read from the input file may be significantly more, or less, than decompressed_size.

The number of decompressed bytes successfully read from the file will be stored in decompressed_read after this function is executed. This value will never be greater than decompressed_size, but it may be less if there was an error or the end of the input file was reached.

Note

*Squash can, and frequently will, read more data from the input file than necessary to produce the requested amount of decompressed data. There is no way to know how much input will be required to produce the requested output, or even how much input was used.

Parameters

file the file to read from

decompressed_size number of bytes to attempt to write to decompressed

decompressed buffer to write the decompressed data to

Returns

The result of the operation

Return values

SQUASH_OK successfully read some data

SQUASH_END_OF_STREAM the end of the file was reached