squash_sys::squash_file_write

pub unsafe extern fn squash_file_write(file: *mut SquashFile, uncompressed_size: size_t, uncompressed: *const uint8_t) -> SquashStatus

Write data to a compressed file.

Attempt to write the compressed equivalent of uncompressed to a file. The number of bytes of compressed data written to the output file may be significantly more, or less, than the uncompressed_size.

Note

It is likely the compressed data will actually be buffered, not immediately written to the file. For codecs which support flushing you can use squash_file_flush to flush the data, otherwise the data may not be written until squash_file_close or squash_file_free is called.

Parameters

file file to write to

uncompressed_size number of bytes of uncompressed data in uncompressed to attempt to write

uncompressed data to write

Returns

Result of the operation