squash_sys::squash_plugin_init

pub unsafe extern fn squash_plugin_init(plugin: *mut SquashPlugin) -> SquashStatus

Load a SquashPlugin

Note

This function is generally only useful inside of a callback passed to squash_foreach_plugin. Every other way to get a plugin (such as [squash_get_plugin]) will initialize the plugin as well (and return NULL instead of the plugin if initialization fails). The foreach functions, however, do not initialize the plugin since doing so requires actually loading the plugin.

Parameters

plugin The plugin to load.

Returns

A status code.

Return values

SQUASH_OK The plugin has been loaded.

SQUASH_UNABLE_TO_LOAD Unable to load plugin.