squash_sys::squash_set_memory_functions

pub unsafe extern fn squash_set_memory_functions(memfn: SquashMemoryFuncs)

Set memory management functions

The aligned_alloc and aligned_free functions may be NULL, as well as either malloc or calloc. Other callbacks require a value.

Note

If you choose to call this function then you must do so before any other function in the Squash, or your program will likely crash (due to attempting to free a buffer allocated with the standard allocator using your non-standard free function).

While Squash itself does not call other memory management functions (such as malloc and free) directly, we can't make any promises about plugins or third-party libraries. We try to make sure as many as possible support custom memory management functions (often filing bugs and patches upstream), but it is unlikely we will ever reach 100% coverage.

Parameters

memfn Functions to use to manage memory