squash_sys::squash_stream_destroy

pub unsafe extern fn squash_stream_destroy(stream: *mut c_void)

Destroy a stream.

Warning

This function must only be used to implement a subclass of SquashObject. Each subclass should implement a *_destroy function which should perform any operations needed to destroy their own data and chain up to the *_destroy function of the base class, eventually invoking squash_object_destroy. Invoking this function in any other context is likely to cause a memory leak or crash. If you are not creating a subclass, you should be calling squash_object_unref instead.

Parameters

stream The stream.