squash_sys::squash_stream_init

pub unsafe extern fn squash_stream_init(stream: *mut c_void, codec: *mut SquashCodec, stream_type: SquashStreamType, options: *mut SquashOptions, destroy_notify: SquashDestroyNotify)

Initialize a stream.

Warning

This function must only be used to implement a subclass of SquashStream. Streams returned by other functions will already be initialized, and you must not call this function on them; doing so will likely trigger a memory leak.

Parameters

stream The stream to initialize.

codec The codec to use.

stream_type The stream type.

options The options.

destroy_notify Function to call to destroy the instance.

See also

squash_object_init