squash_sys::squash_version_api

pub unsafe extern fn squash_version_api() -> *const c_char

Get the API version.

Unlike the library version, the API version will only change when backwards-incompatible changes are made (i.e., it should be "1.0" for every 1.x release). Code linked against a library version not exactly equal to what it was compiled with will almost certainly fail.

Returns

the API version