Struct squash_sys::SquashCodecImpl

pub struct SquashCodecImpl {
    pub info: SquashCodecInfo,
    pub options: *const SquashOptionInfo,
    pub create_stream: Option< extern "C" fn(codec: *mut SquashCodec, stream_type: SquashStreamType, options: *mut SquashOptions) -> *mut SquashStream>,
    pub process_stream: Option< extern "C" fn(stream: *mut SquashStream, operation: SquashOperation) -> SquashStatus>,
    pub splice: Option< extern "C" fn(codec: *mut SquashCodec, options: *mut SquashOptions, stream_type: SquashStreamType, read_cb: SquashReadFunc, write_cb: SquashWriteFunc, user_data: *mut c_void) -> SquashStatus>,
    pub decompress_buffer: Option< extern "C" fn(codec: *mut SquashCodec, decompressed_size: *mut size_t, decompressed: *mut uint8_t, compressed_size: size_t, compressed: *const uint8_t, options: *mut SquashOptions) -> SquashStatus>,
    pub compress_buffer: Option< extern "C" fn(codec: *mut SquashCodec, compressed_size: *mut size_t, compressed: *mut uint8_t, uncompressed_size: size_t, uncompressed: *const uint8_t, options: *mut SquashOptions) -> SquashStatus>,
    pub compress_buffer_unsafe: Option< extern "C" fn(codec: *mut SquashCodec, compressed_size: *mut size_t, compressed: *mut uint8_t, uncompressed_size: size_t, uncompressed: *const uint8_t, options: *mut SquashOptions) -> SquashStatus>,
    pub get_uncompressed_size: Option< extern "C" fn(codec: *mut SquashCodec, compressed_size: size_t, compressed: *const uint8_t) -> size_t>,
    pub get_max_compressed_size: Option< extern "C" fn(codec: *mut SquashCodec, uncompressed_size: size_t) -> size_t>,
    pub _reserved1: Option< extern "C" fn()>,
    pub _reserved2: Option< extern "C" fn()>,
    pub _reserved3: Option< extern "C" fn()>,
    pub _reserved4: Option< extern "C" fn()>,
    pub _reserved5: Option< extern "C" fn()>,
    pub _reserved6: Option< extern "C" fn()>,
    pub _reserved7: Option< extern "C" fn()>,
    pub _reserved8: Option< extern "C" fn()>,
}

Function table for plugins.

This struct should only be used from within a plugin.

This structure may grow over time to accomodate new features, so when setting up the callbacks in a plugin you must set each field individually instead of copying an entire instance of the struct.

Fields

info

Capability information about the codec.

options

options which may bo passed to the codec to modify its operation

create_stream

Create a new SquashStream.

Parameters

codec The codec.

stream_type The type of stream to create.

options The stream options.

Returns

A new SquashStream.

process_stream

Process a SquashStream.

Parameters

stream The stream.

operation The operation to perform.

Returns

A status code.

splice

Splice.

Parameters

options Options to use

stream_type Whether to compress or decompress

read_cb Callback to use to read data

write_cb Callback to use to write data

user_data Date to pass to the callbacks

Returns

A status code

decompress_buffer

Decompress a buffer.

Parameters

codec The codec.

compressed The compressed data.

compressed_size Size of the compressed data.

uncompressed Buffer in which to store the uncompressed data.

uncompressed_size Location of the buffer size on input, used to store the size of the uncompressed data on output.

options Decompression options (or NULL)

compress_buffer

Compress a buffer.

Parameters

codec The codec.

uncompressed The uncompressed data.

uncompressed_size The size of the uncompressed data.

compressed Buffer in which to store the compressed data.

compressed_size Location of the buffer size on input, used to store the size of the compressed data on output.

options Compression options (or NULL)

compress_buffer_unsafe

Compress a buffer.

Plugins implementing this function can be sure that compressed is at least as long as the maximum compressed size for a buffer of uncompressed_size bytes.

Parameters

codec The codec.

uncompressed The uncompressed data.

uncompressed_size The size of the uncompressed data.

compressed Buffer in which to store the compressed data.

compressed_size Location of the buffer size on input, used to store the size of the compressed data on output.

options Compression options (or NULL)

get_uncompressed_size

Get the buffer's uncompressed size.

Parameters

codec The codec.

compressed Compressed data.

compressed_size Size of compressed data (in bytes).

Returns

Size of the uncompressed data, or 0 if unknown.

get_max_compressed_size

Get the maximum compressed size.

Parameters

codec The codec.

uncompressed_size Size of the uncompressed data.

Returns

The maximum buffer size necessary to contain the compressed data.

_reserved1

Reserved for future use.

_reserved2

Reserved for future use.

_reserved3

Reserved for future use.

_reserved4

Reserved for future use.

_reserved5

Reserved for future use.

_reserved6

Reserved for future use.

_reserved7

Reserved for future use.

_reserved8

Reserved for future use.