Struct squash_sys::SquashOptionInfo

pub struct SquashOptionInfo {
    pub name: *const c_char,
    pub _type: SquashOptionType,
    pub info: SquashOptionInfoUnion,
    pub default_value: SquashOptionValue,
}

Information about options which can be passed to a codec.

Fields

name

name of the option

_type

type of the option

info

detailed information about the value

default_value

value to use if none is provided by the user