Struct squash_sys::SquashStream

pub struct SquashStream {
    pub base_object: SquashObject,
    pub _priv: *mut SquashStreamPrivate,
    pub next_in: *const uint8_t,
    pub avail_in: size_t,
    pub total_in: size_t,
    pub next_out: *mut uint8_t,
    pub avail_out: size_t,
    pub total_out: size_t,
    pub codec: *mut SquashCodec,
    pub options: *mut SquashOptions,
    pub stream_type: SquashStreamType,
    pub state: SquashStreamState,
    pub user_data: *mut c_void,
    pub destroy_user_data: SquashDestroyNotify,
}

Compression/decompression streams.

Fields

base_object

Base object.

_priv

Private data.

This is managed internally by Squash and should not be modified by consumers or plugins.

next_in

The next input date to consume.

avail_in

Size (in bytes) of available input.

total_in

The total number of bytes input.

This is managed internally by Squash and should not be modified by consumers or plugins.

next_out

The buffer to write output to.

avail_out

Number of bytes available in the output buffer.

total_out

Total number of bytes output.

This is managed internally by Squash and should not be modified by consumers or plugins.

codec

Codec used for this stream.

options

Options used for this stream.

stream_type

Stream type.

state

State the stream is in.

This is managed internally by Squash and should not be modified by consumers or plugins.

user_data

User data.

Note that this is for consumers of the library, not for plugins. It should be safe to use this from your application.

destroy_user_data

Callback to invoke on user_data when it is no longer necessary.